- suppress reports that are tied to a module if that module is not active; - note that a particular category *could* also be logically tied to a module if, say, that category only contains reports from one module, but I'm not making any effort to hide such a category.
